moved StudentCard component to students folder and adjusted imports
